Users are able to freely remove favourited items by swiping left.
Users are able to remove cart items by swiping as well.
However, i encountered and issue which involves the recycler view. There are 2 activities using the same recycler view thus the app crashes when it doesn't know which views to assign. This is resolved by clearing the view in the cartAdapter class and then updating it notifying the changes to the dataset.

*Resolved*